+++ src/Engine/src/Debug/Instrumentor.h
@@ -46,6 +46,7 @@
 #include <chrono>
 #include <algorithm>
 #include <fstream>
+#include <gsl/gsl>
 
 #include <thread>
 
@@ -160,7 +161,7 @@ namespace BeeEngine
                 auto start = std::chrono::time_point_cast<std::chrono::microseconds>(m_StartTimepoint).time_since_epoch().count();
                 auto end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch().count();
 
-                uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+                uint32_t threadID = gsl::narrow_cast<uint32_t>(std::hash<std::thread::id>{}(std::this_thread::get_id()));
                 Instrumentor::Get().WriteProfile({ m_Name, start, end, threadID });
 
                 m_Stopped = true;
